// packages/diagram/src/LikeC4Diagram.props.d.ts
type Any = t.aux.Any;
type Unknown = t.aux.UnknownLayouted;
type ViewId<A> = t.aux.ViewId<A>;
type Fqn<A> = t.aux.Fqn<A>;
type RelationId = t.aux.RelationId;
type DeploymentFqn<A> = t.aux.DeploymentFqn<A>;
type StrictViewId<A> = t.aux.StrictViewId<A>;
export type NodeRenderers = Partial<CamelCasedProperties<Types.NodeRenderers>>;
export type { WhereOperator };
export type DiagramNodeWithNavigate<A extends Any = Unknown> = SetRequired<DiagramNode<A>, 'navigateTo'>;
// ...
export type OpenSourceParams<A extends Any = Unknown> = {
    element: Fqn<A>;
    property?: string;
} | {
    relation: RelationId;
} | {
    deployment: DeploymentFqn<A>;
    property?: string;
} | {
    view: StrictViewId<A>;
} | {
    view: StrictViewId<A>;
    astPath: string;
};
/**
 * "Go to source" action
 */
export type OnOpenSource<A extends Any = Unknown> = (params: OpenSourceParams<A>) => void;
export type OnNavigateTo<A extends Any = Unknown> = (to: ViewId<A>, event?: ReactMouseEvent, element?: DiagramNodeWithNavigate<A>) => void;
export type OnNodeClick<A extends Any = Unknown> = (node: DiagramNode<A>, event: ReactMouseEvent) => void;
export type OnEdgeClick<A extends Any = Unknown> = (edge: DiagramEdge<A>, event: ReactMouseEvent) => void;
/**
 * On pane/canvas click (not on any node or edge)
 */
export type OnCanvasClick = (event: ReactMouseEvent) => void;
export type OnCanvasContextMenu = (event: ReactMouseEvent) => void;
export type OnInitialized = (params: {
    diagram: DiagramApi;
    xyflow: XYFlowInstance;
}) => void;
export type OnLayoutTypeChange = (layoutType: LayoutType) => void;
export interface LikeC4DiagramEventHandlers<A extends Any = Unknown> {
    onNavigateTo?: OnNavigateTo<A> | null | undefined;
    onNodeClick?: OnNodeClick<A> | null | undefined;
    onNodeContextMenu?: OnNodeClick<A> | null | undefined;
    onCanvasContextMenu?: OnCanvasContextMenu | null | undefined;
    onEdgeClick?: OnEdgeClick<A> | null | undefined;
    onEdgeContextMenu?: OnEdgeClick<A> | null | undefined;
    onCanvasClick?: OnCanvasClick | null | undefined;
    onCanvasDblClick?: OnCanvasClick | null | undefined;
    onLogoClick?: null | undefined | (() => void);
    onOpenSource?: OnOpenSource<A> | null | undefined;
    onInitialized?: OnInitialized | null | undefined;
    /**
     * Triggered when user changes layout type (e.g. from 'manual' to 'auto')
     * Expected another `view` to be passed to the diagram
     *
     * @param layoutType new layout type
     */
    onLayoutTypeChange?: OnLayoutTypeChange | null | undefined;
}